Move to ray.util
