+++ python/ray/exceptions.py
@@ -81,7 +81,7 @@ class RayTaskError(RayError):
         else:
             self.proctitle = setproctitle.getproctitle()
         self.pid = pid or os.getpid()
-        self.ip = ip or ray._private.services.get_node_ip_address()
+        self.ip = ip or ray.util.get_node_ip_address()
         self.function_name = function_name
         self.traceback_str = traceback_str
         # TODO(edoakes): should we handle non-serializable exception objects?
