+++ python/ray/util/__init__.py
@@ -1,3 +1,4 @@
+from ray._private.services import get_node_ip_address
 from ray.util import iter
 from ray.util.actor_pool import ActorPool
 from ray.util.check_serialize import inspect_serializability
@@ -21,6 +22,7 @@ __all__ = [
     "placement_group",
     "placement_group_table",
     "get_placement_group",
+    "get_node_ip_address",
     "remove_placement_group",
     "inspect_serializability",
     "collective",
