+++ rllib/evaluation/rollout_worker.py
@@ -1154,7 +1154,7 @@ class RolloutWorker(ParallelIteratorWorker):
 
     def get_node_ip(self) -> str:
         """Returns the IP address of the current node."""
-        return ray._private.services.get_node_ip_address()
+        return ray.util.get_node_ip_address()
 
     def find_free_port(self) -> int:
         """Finds a free port on the current node."""
